# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Документация к API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/messages": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Отправить сообщение в чат",
                "operationId": "sendMessage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "chat_id",
                                    "temp_id"
                                ],
                                "properties": {
                                    "content": {
                                        "description": "Текст сообщения",
                                        "type": "string",
                                        "maxLength": 65535,
                                        "minLength": 1,
                                        "example": "Привет, как дела?"
                                    },
                                    "chat_id": {
                                        "description": "ID чата, куда отправляется сообщение",
                                        "type": "string",
                                        "minLength": 1,
                                        "example": "chat-123abc"
                                    },
                                    "temp_id": {
                                        "description": "Временный ID сообщения на клиенте для отслеживания статуса",
                                        "type": "string",
                                        "minLength": 1,
                                        "example": "temp-456def"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Сообщение успешно создано, без содержимого в ответе"
                    },
                    "403": {
                        "description": "Доступ запрещён. Пользователь не является участником чата",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You do not have permission to see this message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Ошибки валидации по полям",
                                            "type": "object",
                                            "example": {
                                                "content": [
                                                    "Сообщение не может быть пустым"
                                                ],
                                                "chat_id": [
                                                    "ID чата обязателен"
                                                ],
                                                "temp_id": [
                                                    "Временный ID обязателен"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован - требуется авторизация через Sanctum cookie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/messages/mark_messages_is_read": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Отметить сообщения как прочитанные",
                "operationId": "markMessagesAsRead",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message_ids",
                                    "user_id",
                                    "chat_id"
                                ],
                                "properties": {
                                    "message_ids": {
                                        "description": "Массив ID сообщений для отметки как прочитанных",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "minItems": 1
                                    },
                                    "user_id": {
                                        "description": "ID пользователя, отметившего сообщения",
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 42
                                    },
                                    "chat_id": {
                                        "description": "ID чата, содержащего сообщения",
                                        "type": "string",
                                        "minLength": 1,
                                        "example": "chat-123abc"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Успешно - тело ответа пустое"
                    },
                    "422": {
                        "description": "Ошибка валидации запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Ошибки валидации по полям",
                                            "type": "object",
                                            "example": {
                                                "message_ids": [
                                                    "Массив ID сообщений обязателен и не должен быть пустым"
                                                ],
                                                "user_id": [
                                                    "ID пользователя обязателен"
                                                ],
                                                "chat_id": [
                                                    "ID чата обязателен"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован - требуется авторизация через Sanctum cookie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Sanctum cookie-based auth. Требуется авторизация через web, после чего браузер отправляет куки: laravel_session (сессия) и XSRF-TOKEN (csrf).",
                "name": "XSRF-TOKEN",
                "in": "cookie"
            }
        }
    },
...
}